feat: Allow for redirection of config2json output
Currently in xpctl the `config2json` command requires a `filename`
argument for where to save the resulting configuration. This means that
in order to examine/query the output config in the terminal we need to do something
like `xpctl config2json ${TASK} ${SHA1} /tmp/a.json && cat /cat/a.json |
jq` Given that this is one of the most common things I do with
xpctl (checking what the config is for some results) it was very
annoying to do each time.

This PR updates the `config2json` command to be more unix like and allow
for output redirection. The `filename` is now an optional argument, when
it is given it the code will work the exact same. When the file name is
not given the it echo the the result to the screen so that we can do
things like redirection to a file or pipe it into something like `jq`.

_note_ unlike a normal unix tool when output the json result we don't
output it as a single line. This could be weird but it given the `click`
integration I don't know how easy it is to detect if the output is the
terminal (were we would print indented json) and then we are
redirected (and would output a single line). There is also a question of
if we should do different behavior on if it is a file or a pipe (indent
the first, single line the latter) and if this is detectable.

It also updates some of the logging so it goes to `stderr` instead of
`stdout` which was breaking the redirection.